change to double for gradient checks

+++ dl4j-examples/src/main/java/org/deeplearning4j/examples/advanced/features/customizingdl4j/lossfunctions/CustomLossUsageEx.java
@@ -102,12 +102,12 @@ public class CustomLossUsageEx {
     }
 
     private static DataSetIterator getTrainingData(int batchSize, Random rand){
-        double [] sum = new double[nSamples];
-        double [] input1 = new double[nSamples];
-        double [] input2 = new double[nSamples];
+        float [] sum = new float[nSamples];
+        float [] input1 = new float[nSamples];
+        float [] input2 = new float[nSamples];
         for (int i= 0; i< nSamples; i++) {
-            input1[i] = MIN_RANGE + (MAX_RANGE - MIN_RANGE) * rand.nextDouble();
-            input2[i] =  MIN_RANGE + (MAX_RANGE - MIN_RANGE) * rand.nextDouble();
+            input1[i] = MIN_RANGE + (MAX_RANGE - MIN_RANGE) * rand.nextFloat();
+            input2[i] =  MIN_RANGE + (MAX_RANGE - MIN_RANGE) * rand.nextFloat();
             sum[i] = input1[i] + input2[i];
         }
         INDArray inputNDArray1 = Nd4j.create(input1, new int[]{nSamples,1});
